Add getKeyPairBySpki to platformKeys
Add the new method chrome.platformKeys.getKeyPairBySpki() that,
similarly to chrome.platformKeys.getKeyPair() returns subtleCrypto
objects for a private and a public key.
The key pair already exists on the device, and getKeyPairBySpki()
receives a public_key_spki_der as input to identify which key pair
to return.
The already existing API function getKeyPair() receives a certificate,
extracting the public_key_spki_der from it and returning the key
pair based on that spki.
In our use case (the Imprivata extension) we don't have access to a
certificate in order for us to use getKeyPair(). A key pair is needed
in Imprivata for sending authorization requests (signing data in the
request with the private key) to the Imprivata server, which verifies
our authorization request using the public key.

Bug: 1073466
Change-Id: I44ef0335ede498b0e5f4878554050999dbbeda85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190675
Commit-Queue: Maria Petrisor <mpetrisor@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Omar Morsi <omorsi@google.com>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781300}