[iOS Enterprise] Refactor BrowserDMTokenStorage to delegate pattern
In preparation for moving BrowserDMTokenStorage to components, this
change moves platform-specific DM token management code to delegates
instead of using derived classes. In this case, the derived class
approach won't work after the move to components because the current
code has the derived classes implementing the parent's Get() singleton
method, which won't work when the base and the derived classes are in
different DLLs.

Summary of the changes:
- Pure virtual methods in BrowserDMTokenStorage have been moved to a
  delegate inner class.
- Platform-specific token management classes now derive from the
  delegate instead of the BrowserDMTokenStorage singleton.
- CBCMController is responsible for creating the correct platform
  delegate and assigning it to the BrowserDMTokenStorage singleton
  during early initialization.
- For tests, this refactor is completely transparent, as the
  FakeBrowserDMTokenStorage creates and manages its own fake delegate
  implementation.


Bug: 1091929
Change-Id: I02ae185cc5cbde46016039b49cefaeb8e963ec4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246979
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781341}