Implement page-based scrolling in compositor
Page based scrolling is used in three situations:

1. PageUp/PageDown keys
2. Main thread scrollbar scrolling using the track
3. Wheel when a scroll-by-page setting is enabled

The first two are explicitly main thread only. The last use is a niche
use case and was simply never implemented in the compositor. With recent
additions related to percent based scrolling this becomes trivial to
handle so this CL allows LayerTreeHostImpl to handle scrolls using a
page granularity.

Bug: 1082589
Change-Id: I35fc7c4fa3c57ba2c8c86c89fed314c516f8b4f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241131
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Rahul Arakeri <arakeri@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#781405}