Disable Direct Composition or overlay when the function fails to prevent a blank screen.
Returning a failure in the Direct Composition functions will trigger a
context lost in GL Renderer and Direct Composition will be reinitialized.
If the same Direct Composition function keeps failing, it will cause an
infinity loop of restart and the result is a blank screen or a blank
video.

If a DXGI or D3D11 call fails due to out of memory or a bad graphics
driver, most likely the same call will fail next time. We should disable
Direct Composition or Overlay to prevent the same failure.

Bug: 1091958
Change-Id: I3ac7d9bac0fdf5938302583b072e7bda5839d2c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247390
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781408}