Remove WebTestControlHost flush to renderer main thread.
For crbug.com/889952 a flush was introduced to get messages sent to the
renderer process compositor thread to be handled (and bounced to the
main thread) before navigating and starting the test on the renderer
main thread.

A second flush was also added at that time to the main thread, which
would ensure those bounced compositor-thread things would be handled
there before proceeding. But since those events don't have side effects
that need to be received by the browser before proceeding, this second
flush is not needed. The tasks for compositor-thread events bounced to
the main thread will be run before the main thread navigates and starts
running the test regardless.

The fast/selectors/selection-window-inactive.html test was disabled as
flaky for bug 889952, but the flushes added to that test did not
actually stop it from being flaky.

R=avi@chromium.org

Bug: 889952, 953991, 1084377
Change-Id: Iea96b364efd36d057f156f940aee19a64aaacf39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261156
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781473}