[LargestContentfulPaint] Fix edge case in experimental LCP
This bug fixes a problem that may arise when the largest image is
loading and then removed. Before, we'd continue to wait for the paint
time of such image, and hence would only correctly update the
experimental LCP value if we ever painted something at least as large as
that removed image.

Bug: 1045640
Change-Id: I356a58215003871cc15e3bee3ed86423ac612934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255498
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781502}