Carry text-decoration thickness in applied text decorations
On each decoration level, a thickness can be specified that is
individual to the decoration specified at this decorating box /
level. Carry those in AppliedTextDecorations to ensure that ancestral
decorations do not get incorrectly painted using the stroke thickness
that is specified on the current element.

A previous attempt had been made to carry the resolved thickness (e.g. a
percentage resolved against a font size but that won't work due to the
AppliedTextDecoration vector being const in ComputedStyle and
ComputedStyle generally kept const). Also, computing the resolved
thickness in ComputeDecorationInfo fits the current conventions better.

Thus, instead of carrying a single thickness in DecorationInfo, which
overrides all thicknesses of previous AppliedTextDecorations' carry a
parallel array with as many entries as there are AppliedTextDecorations
and store the resolved thicknesses in those.

Bug: 1096038
Change-Id: Ifdaf6dd167024a919fafae297e5a01a226b9cda1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252081
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781528}