Trust Tokens: Allow nonexistent directories in database paths
In some circumstances, ProfileNetworkContextService can set the
directories it populates in NetworkContextParams (for the cookie store,
reporting store, and so on) to a nonexistent path. Currently, the Trust
Tokens backing store doesn't attempt to create directories when
nonexistent, causing some tests to fail flakily when sql::Database::Open
hits a LOG(DFATAL) after attempting to open a database at a path with a
nonexistent parent directory. (It's flaky because the failure races with
other backing stores that do attempt to create the directory in this
scenario.)

This change hopefully resolves the failures, by making
SQLiteTrustTokenPersister::CreateForFilePath create the given path's
directory if it does not exist.

R=mmenke

Fixed: 1098019
Change-Id: Id075718035e1fd7e29cebc063d9126f1f33c4c83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258685
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781534}