[Paint Preview] Capture and restore scroll state
In order to give a more seamless transition. This CL captures the scroll
state of the page at time of recording. Then this can be used to replay
the scroll position when loading the paint preview. This behavior can
be disabled via PlayerManager if not desired.

Bug: 1096598
Change-Id: I4dffa033e12d5cac9cbf521a330cefa9201b4ba7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250589
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781540}