Remove SecondaryTestWindowObserver
This drops one dependency from shell/browser/ to shell/browser/web_test
by going directly to WebTestControlHost and having it register a
WebContentsObserver to hear about new RenderFrames.

This still has Shell using WebTestControlHost, but we're narrowing scope
down to just that one class, at which point we can insert a delegate
instead of the concrete type.

We remove the WebTestControlHost::current_pid_ and the notification
code around it, which is was partly tied to RenderFrame creation
notification and was both wholly unnecessary as well as problematic or
incorrect if the main frame moved to a new renderer process on
navigation during a test.

R=avi@chromium.org

Bug: 866140, 1069111
Change-Id: I9e0622fa9ec73999bf59f2ab794acfd4684052d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247090
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781553}