Fix crash in GraphicsDelegateWin when the GL context generates an error
GraphicsDelegateWin creates GL resources in PreRender and deletes them
in PostRender. If errors are generated in PreRender, PostRender still
tries to delete them, which fires a DCHECK.

This change returns false from PreRender if there are any errors in
the GL context. In this case, the current overlay pose is ignored in
VRBrowserRendererThreadWin.

Bug: 1090957
Change-Id: I7aa4d3b1b0eb26ad571d9de396f3ef2a569c2382
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2259232
Commit-Queue: Patrick To <patrto@microsoft.com>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781572}