cvox: Send right click when hovering over anchor point in lift activated region.
For lift activated regions like the virtual keyboard, it is quite
difficult to perform a long press, because lifting the finger to perform
passthrough will immediately trigger a press.

A solution to this is to send a right click event when the user is
hovering over the same anchor point for 5s in a lift activated region.

Bug: b/157374150
Change-Id: I93ac8e02e45a9a11a9e846504ea6320cd7e3f592
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224474
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781622}