[consent toggle] Fix logic for incrementing shown count.
There is a bug where the Suggested Content shown count could be
incremented even if it was not shown in the launcher.

MaybeIncreaseSuggestedContentInfoShownCount currently depends on the
value of ShouldShowAssistantPrivacyInfo. Since the Assistant count is
incremented first, the value of ShouldShowAssistantPrivacyInfo could
change and be outdated by the time Suggested Content does its own
increment.

This change fixes the logic so that Suggested Content shown count is
not incremented if Assistant is incremented.

Bug: 1079169
Change-Id: Ib632312504ce4d5102d47355fef2e1dbfe81cbc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2259694
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Commit-Queue: Rachel Wong <wrong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781625}