IsolatedPrerender Correctly Handles Cookies
IsolatedPrerender handles cookies in the following way:
* Cookies which are set on the prefetched mainframe are kept in the
  isolated NetworkContext's cookie jar.
* Subresources fetched during NoStatePrefetch use the isolated
  NetworkContext's cookie jar and so send cookies as appropriate.
* When the user navigates to a link that was prefetched, any cookies
  for that page are copied over to the normal profile.
* Subresources that were not prefetched before the click use the
  prefetch cookie when they are made after the click with the normal
  profile.

Note that copying the cookies over to the normal profile is a multi-
step process during which the prefetched mainframe is blocked from
committing to ensure page correctness.

Important: Opaque IsolationInfo's block all cookies so this CL also
transitions IsolatedPrerenders to a normal navigation-style NIK and
IsolationInfo. This is ok because all prefetches are occuring in a
separate NetworkContext from the rest of the browser and therefore
were never at risk of having shared resources.

Bug: 1096765
Change-Id: I090fc57d46936b4972bd799eab3e4b022019bbc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253200
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781703}