[WebLayer] Reset infobar container visibility on navigation
This CL has InfoBarContainer reset itself to being visible when a
navigation commits. This is an adaptation of Chrome's behavior: in
Chrome some features hide the infobar container, which should always be
undone on navigation. While no feature in WebLayer yet does this, it is
good to have this defensive behavior in place so that any such added
features don't end up inadvertently hiding the infobar container
"forever" in a given tab.

Bug: 1093846
Change-Id: I9ead10a1be0b670d8b3f14c195657e60a4a43f20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254022
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781743}