[Nearby] Add empty chrome://nearby WebUI
This will host the Nearby Share UI for Chrome browser and ChromeOS. It
renders an example Polymer component that will be embedded in various
placed depending on the OS. CrOS will show it in a WebView in the native
Share Sheet and other platforms will have a standalone dialog. The page
is guarded by the kNearbySharing feature flag.

Bug: 1079271
Change-Id: Ia72ef02876518bd6dd594fa2b0fa88fbb8f4d3e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187614
Commit-Queue: Richard Knoll <knollr@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: James Vecore <vecore@google.com>
Cr-Commit-Position: refs/heads/master@{#781752}