Add an helper function to DidNavigateFrame in unittests.
When experimenting with COOP access reporting. I have seen myself
modifying the signature of RenderFrameHostManager and updating over and
over the associated unit tests. I lost a lot of time on this. This patch
will prevent me from continuing.

This patch replaces all the call to this function by an helper function.
The helper function provides all the necessary arguments that aren't
tested and are constant throughout the whole file.

Bug: chromium:1090273
Change-Id: Ia0dc4b6cb1e9ab0b0d209dbfcf940b2295ebb7b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260695
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781755}