X11 and Ozone: //content: choose code path with IsUsingOzonePlatform
This patch adds for usage of ozone platform in //content.

It's not possible to use the linux wrapper [1] for everything as it
may introduce circular deps and other problems. Plus, there
are many places that are guarded either by single a single USE_X11
or USE_OZONE. Thus, use the IsUsingOzonePlatform feature
flag to choose the correct path.

[1] https://crrev.com/c/2250106

-----

Please note that this ugliness is added temporarily and
will be removed as soon as use_x11 is removed (hopefully
by Q1 2021 depending on how the finch trial goes).

Please also note that it's impossible to build use_x11 && use_ozone
without some hacks in PlatformCursor code. The changes to that are
on their way to upstream.

----

Bug: 1085700
Change-Id: I51054a82bcad34f3f887379003cf24a66a1e9928
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250110
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#781758}