BackForwardCache: Fix issues around first-paint after restoring
Before this CL, the detection whether the page was in background between
when the page was restored the first-paint happened was not accurate.
The checking function checked only the last timings were checked and did
not consider multiple restorings. This CL fixes this issue by recording
all the backgrounding states related to restoring, and check an
appropriate state by specifying an index.

This CL also added an item to histograms.xml

Change-Id: I9ae6b0bb8ee192f21ae42c518166372106bfb878
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251526
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781770}