[WebLayer] Handle accessibility in infobars
This CL augments WebLayer's infobars implementation by handling
accessibility as Chrome's does: while accessibility is enabled the
infobar container view does not scroll out of view. The implementation
is a port of that of //chrome's InfoBarContainer.java extended to
handle the wrinkle that WebLayerAccessibilityUtil drops its observers
when there are no more Browsers and doesn't re-add them on addition of
a new Browser.

Bug: 1093846
Change-Id: Ic58d17823068d60c9d0d25a80f7ddcf420f9a2fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254020
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781801}